requirements.txt
python-Levenshtein
requests
transmissionrpc
tabulate
python-qbittorrent
